+++ embedded-hal-bus/Cargo.toml
@@ -15,11 +15,13 @@ version = "0.1.0-alpha.3"
 
 [features]
 std = []
+async = ["dep:embedded-hal-async"]
 
 [dependencies]
 embedded-hal = { version = "=1.0.0-alpha.11", path = "../embedded-hal" }
+embedded-hal-async = { version = "=0.2.0-alpha.2", path = "../embedded-hal-async", optional = true }
 critical-section = { version = "1.0" }
 
 [package.metadata.docs.rs]
-features = ["std"]
+features = ["std", "async"]
 rustdoc-args = ["--cfg", "docsrs"]
