+++ embedded-hal-bus/src/asynch/mod.rs
@@ -1,0 +1,3 @@
+//! Bus sharing utilities for [`embedded-hal-async`]
+
+pub mod spi;
