+++ embedded-hal-bus/src/lib.rs
@@ -2,6 +2,11 @@
 #![warn(missing_docs)]
 #![cfg_attr(not(feature = "std"), no_std)]
 #![cfg_attr(docsrs, feature(doc_cfg))]
+#![cfg_attr(feature = "async", feature(async_fn_in_trait, impl_trait_projections))]
 
 pub mod i2c;
 pub mod spi;
+
+#[cfg(feature = "async")]
+#[cfg_attr(docsrs, doc(cfg(feature = "async")))]
+pub mod asynch;
